addhasilmarcopush2

+++ routes/drivers.js
@@ -4,26 +4,42 @@ const Models = require('../models')
 const routes = express.Router();
 
 routes.get('/', function(req,res) {
-    Models.Driver.findAll()
-    .then(drivers => {
-        // res.send(drivers);
-
-        let obj = {
-            title: 'List Drivers OmRia.com',
-            arrDrivers: drivers
-        }
-        res.render('drivers/driver_list.ejs', obj)
-    })
+    // res.send(req.session);
+    let sessionData = req.session;
+    // console.log(sessionData);
+    if(sessionData.driver) {
+        Models.Driver.findAll({where:{id:sessionData.driver.id}})
+        .then(driver => {
+            // res.send(driver);
+    
+            let obj = {
+                title: 'Profile Drivers OmRia.com',
+                arrDrivers: driver,
+                isLogin: sessionData
+            }
+            res.render('drivers/driver_list.ejs', obj)
+        })    
+    } else {
+        res.redirect('/login/driver');
+    }
 })
 
-
 routes.get('/add', function(req,res) {
-    let obj = {
+    
+    let sessionData = req.session;
+    
+    if(sessionData.driver) {
+            let obj = {
         title: 'New Driver OmRia.com Form',
         error: req.query.err
     }
     // console.log(obj.error)
     res.render('drivers/driver_add.ejs', obj)
+    } else {
+        res.redirect('/login/driver');
+    }
+
+
 })
 
 
@@ -36,7 +52,8 @@ routes.post('/add', function(req,res) {
         Email: data.Email,
         Phone: data.Phone,
         Gender: data.Gender,
-        LicensePlate: data.LicensePlate
+        LicensePlate: data.LicensePlate,
+        Password: data.Password
     }
 
     Models.Driver.create(obj)
@@ -51,17 +68,23 @@ routes.post('/add', function(req,res) {
 
 
 routes.get('/edit/:id', function(req,res) {
-    let id = req.params.id;
+    let sessionData = req.session;
+    
+    if(sessionData.driver) {
+        let id = req.params.id;
 
-    Models.Driver.findById(id)
-    .then(driver => {
-        let obj = {
-            title: 'Edit Driver OmRia.com Form',
-            driver: driver,
-            error: req.query.err
-        }
-        res.render('drivers/driver_edit.ejs', obj)
-    })
+        Models.Driver.findById(id)
+        .then(driver => {
+            let obj = {
+                title: 'Edit Driver OmRia.com Form',
+                driver: driver,
+                error: req.query.err
+            }
+            res.render('drivers/driver_edit.ejs', obj)
+        })
+    } else {
+        res.redirect('/login/driver');
+    }   
 })
 
 
@@ -76,7 +99,8 @@ routes.post('/edit/:id', function(req,res) {
         Email: data.Email,
         Phone: data.Phone,
         Gender: data.Gender,
-        LicensePlate: data.LicensePlate
+        LicensePlate: data.LicensePlate,
+        Password: data.Password
     }
 
     Models.Driver.findById(id)
@@ -103,12 +127,20 @@ routes.post('/edit/:id', function(req,res) {
 })
 
 routes.get('/delete/:id', function(req,res) {
-    let id = req.params.id
 
-    Models.Driver.destroy({where: {id:id}})
-    .then(driver => {
-        res.redirect('/drivers');
-    })
+    let sessionData = req.session;
+    
+    if(sessionData.driver) {
+        let id = req.params.id
+
+        Models.Driver.destroy({where: {id:id}})
+        .then(driver => {
+            res.redirect('/drivers');
+        })
+            
+    } else {
+        res.redirect('/login/driver');
+    }
 })
